backend/chunking-rag-py/requirements.txt
# Web 框架
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18

# LLM 客户端
openai==1.57.0

# Embedding & Rerank
FlagEmbedding==1.3.4
# transformers 5.x removed is_torch_fx_available; FlagEmbedding 1.3.4 requires 4.x
transformers==4.46.3

# 检索
rank-bm25==0.2.2
jieba==0.42.1

# 数值计算（FlagEmbedding 对 numpy 2.x 不完全兼容）
numpy==1.26.4

# 文档解析
PyMuPDF==1.25.1
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5

# 测试
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
